pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "def-mineinsight"
version = "0.1.0"
description = "ANIMA DEF-mineinsight: multi-modal landmine detection on MineInsight dataset"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "AIFLOW LABS" }]
dependencies = [
  "numpy>=1.26",
  "pillow>=10.0",
  "pyyaml>=6.0",
  "torch>=2.2",
  "torchvision>=0.17",
  "opencv-python-headless>=4.9",
  "scipy>=1.12",
  "tqdm>=4.66",
  "tensorboard>=2.16",
  "safetensors>=0.4",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
  "ruff>=0.4",
]

[project.scripts]
mineinsight-train = "mineinsight.train:main"
mineinsight-eval = "mineinsight.evaluate:main"
mineinsight-serve = "mineinsight.serve:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mineinsight"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "C4"]
ignore = ["N812", "B008"]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu128"]
